RAG_Project/requirements.txt
langchain>=1.2.0
langchain-aws>=1.2.0
langchain-community>=0.4.0
langchain-core>=1.2.0
langchain-text-splitters>=1.1.0
faiss-cpu>=1.9.0
pypdf>=5.0.0
boto3>=1.35.0
streamlit>=1.40.0
python-dotenv>=1.0.0
